Rename folder sys_arch_src to driver
- sys_arch_src folder hosts PCIe/SMMU/GIC initialization
  and access interface.
- Rename the sys_arch_src folder to driver folder
- Renamed the paths in the respective test files
