+++ val/driver/gic/v3/AArch64/v3_asm.S
@@ -17,7 +17,7 @@
 
 #ifndef TARGET_EMULATION
 #include "include/acs_pe.h"
-#include "sys_arch_src/gic/acs_exception.h"
+#include "driver/gic/acs_exception.h"
 #else
 #define ICC_IAR1_EL1    S3_0_C12_C12_0
 #define ICC_EOIR1_EL1   S3_0_C12_C12_1
