+++ val/src/acs_test_infra.c
@@ -18,7 +18,7 @@
 #include "include/acs_val.h"
 #include "include/acs_pe.h"
 #include "include/acs_common.h"
-#include "sys_arch_src/gic/acs_exception.h"
+#include "driver/gic/acs_exception.h"
 #include "include/pal_interface.h"
 #include "include/val_interface.h"
